# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Blue Media social platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
  /** @nullable */
  message?: string | null;
}

export interface RegisterInput {
  email: string;
  name: string;
  /**
     * @minLength 4
     * @maxLength 4
     */
  pin: string;
}

export interface LoginInput {
  email: string;
  pin: string;
}

export interface ChangePasswordInput {
  currentPin: string;
  /**
     * @minLength 4
     * @maxLength 4
     */
  newPin: string;
}

export type UserPrivacy = typeof UserPrivacy[keyof typeof UserPrivacy];


export const UserPrivacy = {
  public: 'public',
  friends: 'friends',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  profilePicture?: string | null;
  /** @nullable */
  coverPicture?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  website?: string | null;
  privacy?: UserPrivacy;
  isAdmin?: boolean;
  isBlueAI?: boolean;
  blueBadge?: boolean;
  /** @nullable */
  blueBadgeClaimedAt?: string | null;
  restricted?: boolean;
  banned?: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type UserUpdatePrivacy = typeof UserUpdatePrivacy[keyof typeof UserUpdatePrivacy];


export const UserUpdatePrivacy = {
  public: 'public',
  friends: 'friends',
} as const;

export interface UserUpdate {
  name?: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  website?: string | null;
  privacy?: UserUpdatePrivacy;
}

export interface UserStats {
  friendCount: number;
  postCount: number;
  followerCount: number;
  followingCount: number;
}

export interface FollowStatus {
  isFollowing: boolean;
  followerCount: number;
  followingCount: number;
}

export interface BadgeClaim {
  ok: boolean;
  blueBadge: boolean;
  /** @nullable */
  expiresAt?: string | null;
  message?: string;
}

export interface UploadResult {
  url: string;
}

export type FriendRequestStatus = typeof FriendRequestStatus[keyof typeof FriendRequestStatus];


export const FriendRequestStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export interface FriendRequest {
  id: number;
  requesterId: number;
  addresseeId: number;
  status: FriendRequestStatus;
  requester?: User;
  addressee?: User;
  createdAt: string;
}

export interface FriendRequestInput {
  addresseeId: number;
}

export type FriendshipStatusStatus = typeof FriendshipStatusStatus[keyof typeof FriendshipStatusStatus];


export const FriendshipStatusStatus = {
  none: 'none',
  pending_sent: 'pending_sent',
  pending_received: 'pending_received',
  friends: 'friends',
} as const;

export interface FriendshipStatus {
  status: FriendshipStatusStatus;
  /** @nullable */
  requestId?: number | null;
}

export interface PostReaction {
  type: string;
  count: number;
}

export interface Post {
  id: number;
  userId: number;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  bgColor?: string | null;
  author?: User;
  reactions?: PostReaction[];
  commentCount?: number;
  /** @nullable */
  myReaction?: string | null;
  createdAt: string;
}

export interface PostInput {
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  bgColor?: string | null;
}

export type ReactionInputType = typeof ReactionInputType[keyof typeof ReactionInputType];


export const ReactionInputType = {
  heart: 'heart',
  cry: 'cry',
  laugh: 'laugh',
  angry: 'angry',
} as const;

export interface ReactionInput {
  type: ReactionInputType;
}

export interface Comment {
  id: number;
  postId: number;
  userId: number;
  content: string;
  author?: User;
  createdAt: string;
}

export interface CommentInput {
  content: string;
}

export type ReportInputReason = typeof ReportInputReason[keyof typeof ReportInputReason];


export const ReportInputReason = {
  sexual_content: 'sexual_content',
  harassment: 'harassment',
  hate_speech: 'hate_speech',
  violence: 'violence',
  spam: 'spam',
} as const;

export interface ReportInput {
  reason: ReportInputReason;
}

export type ReportStatus = typeof ReportStatus[keyof typeof ReportStatus];


export const ReportStatus = {
  pending: 'pending',
  resolved: 'resolved',
  dismissed: 'dismissed',
} as const;

export interface Report {
  id: number;
  reporterId: number;
  /** @nullable */
  reportedPostId?: number | null;
  /** @nullable */
  reportedUserId?: number | null;
  reason: string;
  status: ReportStatus;
  createdAt: string;
  reporter?: User;
  reportedUser?: User;
  reportedPost?: Post;
}

export type AdminActionAction = typeof AdminActionAction[keyof typeof AdminActionAction];


export const AdminActionAction = {
  remove_post: 'remove_post',
  restrict_account: 'restrict_account',
  remove_account: 'remove_account',
  ban_account: 'ban_account',
  dismiss_report: 'dismiss_report',
  restore_account: 'restore_account',
} as const;

export interface AdminAction {
  action: AdminActionAction;
}

export type ConversationType = typeof ConversationType[keyof typeof ConversationType];


export const ConversationType = {
  direct: 'direct',
  group: 'group',
} as const;

export interface MessageReaction {
  id: number;
  messageId: number;
  userId: number;
  emoji: string;
  user?: User;
}

export interface Message {
  id: number;
  conversationId: number;
  senderId: number;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  sender?: User;
  reactions?: MessageReaction[];
  seenBy?: number[];
  createdAt: string;
}

export interface Conversation {
  id: number;
  type: ConversationType;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  pictureUrl?: string | null;
  /** @nullable */
  backgroundTheme?: string | null;
  participants?: User[];
  lastMessage?: Message;
  unreadCount?: number;
  createdAt: string;
}

export interface ConversationInput {
  participantId: number;
}

export interface GroupInput {
  name: string;
  /** @nullable */
  pictureUrl?: string | null;
  participantIds: number[];
}

export interface ConversationUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  pictureUrl?: string | null;
  /** @nullable */
  backgroundTheme?: string | null;
}

export interface MessageInput {
  content: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface MessageReactionInput {
  emoji: string;
}

export interface ParticipantInput {
  userId: number;
}

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  friend_request: 'friend_request',
  friend_accepted: 'friend_accepted',
  post_reaction: 'post_reaction',
  post_comment: 'post_comment',
  message: 'message',
  blue_badge: 'blue_badge',
  report_received: 'report_received',
  follow: 'follow',
} as const;

export interface Notification {
  id: number;
  userId: number;
  type: NotificationType;
  /** @nullable */
  fromUserId?: number | null;
  fromUser?: User;
  /** @nullable */
  postId?: number | null;
  /** @nullable */
  conversationId?: number | null;
  /** @nullable */
  message?: string | null;
  read: boolean;
  createdAt: string;
}

export interface UnreadCount {
  count: number;
}

export type SearchUsersParams = {
q?: string;
};

export type ListFriendsParams = {
userId?: number;
};

export type ListPostsParams = {
userId?: number;
};

export type GetAdminReportsParams = {
status?: string;
};