# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Blue Media social platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserPrivacy } from './userPrivacy';

export interface User {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  profilePicture?: string | null;
  /** @nullable */
  coverPicture?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  website?: string | null;
  privacy?: UserPrivacy;
  isAdmin?: boolean;
  isBlueAI?: boolean;
  blueBadge?: boolean;
  /** @nullable */
  blueBadgeClaimedAt?: Date | null;
  restricted?: boolean;
  banned?: boolean;
  createdAt: Date;
}